refactor: CardInfoForm 하위 디렉토리 제거
